updated according to project

+++ temp_monitor/delete.php
@@ -1,11 +1,18 @@
 <?php
-
+session_start();
 include "db.php";
-include "header.php";
-$id = $_GET['id'];
 
-mysqli_query($conn,"DELETE FROM temperature_humidity WHERE id='$id'");
+if (!isset($_SESSION['user_id'])) {
+    header("Location: login.php");
+    exit();
+}
 
-header("Location:index.php");
+$id = intval($_GET['id'] ?? 0);
 
+if ($id > 0) {
+    mysqli_query($conn, "DELETE FROM temperature_humidity WHERE id='$id'");
+}
+
+header("Location:view.php");
+exit();
 ?>
